Added re-routing and error checking when otp and email don't exist in session storage. Fixed some things.
Added re-routing when otp and email don't exist in session storage. Fixed some things. Added a validation to inform user to head back and re-enter. Also if the OTP expires and it is still in the session cache, it displays error, jus tlike if someone changes the email or the otp while being in the second step of the page. NEED TO CONNECT THE COMPONENTS AND CREATE THE SERVICE FOR THE EMAIL SENDER